Add files via upload
This program copies everything which you copy in your computer. It then creates a textfile which all the history is stored. This can be anything in the form of text. It is an infinite loop, that doesn't eat up a lot of memory.

I hope you like it. Don't use it for malicious acts.

With great power comes great responsibility.